Automatisches Abrufen der TAN-Medienbezeichnungen (zusammen mit den SEPA-Infos der Konten) und Bereitstellen dieser im HBCI-Callback NEED_PT_TANMEDIA (identisches Verfahren zu Callback NEED_PT_SECMECH). Damit kann HBCI4Java jetzt die Liste der TAN-Medienbezeichnungen bequem dem Aufrufer im Callback anbieten, ohne dass dieser sie manuell abrufen muss.

+++ src/org/kapott/hbci/passport/HBCIPassportPinTan.java
@@ -252,9 +252,12 @@ public class HBCIPassportPinTan
             HBCIUtils.log("closing output stream", HBCIUtils.LOG_DEBUG);
             o.close();
             
-            HBCIUtils.log("deleting old passport file " + passportfile, HBCIUtils.LOG_DEBUG);
-            if (!passportfile.delete())
-                HBCIUtils.log("delete method for " + passportfile + " returned false", HBCIUtils.LOG_ERR);
+            if (passportfile.exists()) // Nur loeschen, wenn es ueberhaupt existiert
+            {
+                HBCIUtils.log("deleting old passport file " + passportfile, HBCIUtils.LOG_DEBUG);
+                if (!passportfile.delete())
+                    HBCIUtils.log("delete method for " + passportfile + " returned false", HBCIUtils.LOG_ERR);
+            }
 
             // Wenn die Datei noch existiert, warten wir noch etwas
             int retry = 0;
